refactor(llm-router): extract catalog handler factories into catalog/handlers.rs
