+++ llm-router/src/catalog/mod.rs
@@ -1,3 +1,4 @@
+pub mod handlers;
 pub mod queries;
 pub mod reconcile;
 pub mod store;
